Refactor redirect. WIP! #206
